4.7: Implement caching strategy
- Move Real-time channel subscription to conversations layout
- Update logError function

Fixes #54
