+++ src/utils/errorHandler.ts
@@ -2,9 +2,7 @@ import { FetchError } from "ofetch";
 
 export const logError = (error: any) => {
 	if (error instanceof FetchError) {
-		console.error(error.statusCode);
-		console.error(error.statusMessage);
-		console.error(error.stack);
+		console.error(`[${error.statusCode}] ${error.statusMessage}\nStack: ${error.stack}`)
 	} else {
 		console.error(error);
 	}
